Profiles should be created when the 3rd party reference does not exist in the local database - Better if a message can be displayed to notify the agent that the profile was created successfully.
Profiles should not be created if more than one 3rd party reference is received - Should notify the agent that the profile was not created in facetone.

Profiles should be updated if 3rd party reference exists in the local database

Profiles should not be updated if multiple 3rd party references are received - An error message should be
displayed when a profile is not created.

Profiles should not be created when a searched profile with a new 3rd party reference has an existing primary phone number

Profiles should not be updated when a searched profile with an existing 3rd party reference has an existing primary phone number

When 3rd party reference is not sent successfully from the 3rd party service, A profile should not be created. - Developer testing done. Cannot check unless a new method to the dummy service is developed. Effort is time consuming hence not checking the test case.